+++ static/app/utils/url/useQueryParamState.tsx
@@ -3,8 +3,8 @@ import * as Sentry from '@sentry/react';
 
 import {defined} from 'sentry/utils';
 import {type decodeList, decodeScalar, type decodeSorts} from 'sentry/utils/queryString';
+import {useUrlBatchContext} from 'sentry/utils/url/urlParamBatchContext';
 import useLocationQuery from 'sentry/utils/url/useLocationQuery';
-import {useUrlBatchContext} from 'sentry/views/dashboards/widgetBuilder/contexts/urlParamBatchContext';
 
 interface UseQueryParamStateWithScalarDecoder<T> {
   fieldName: string;
